refactor take 1
