+++ energizer/loops/pool_loop.py
@@ -1,0 +1,71 @@
+import os
+from typing import Any, List, Union
+
+from pytorch_lightning.loops.dataloader.evaluation_loop import EvaluationLoop
+from pytorch_lightning.utilities.types import EPOCH_OUTPUT
+
+from energizer.loops.pool_epoch_loop import PoolEvaluationEpochLoop
+
+
+class PoolEvaluationLoop(EvaluationLoop):
+    """Loops over all dataloaders for evaluation."""
+
+    def __init__(self) -> None:
+        super().__init__(verbose=False)
+        self.epoch_loop = PoolEvaluationEpochLoop()
+
+    def _on_evaluation_start(self, *args: Any, **kwargs: Any) -> None:
+        """Runs ``on_pool_start`` hooks."""
+        assert self._results is not None
+        self._results.to(device=self.trainer.lightning_module.device)
+
+        self.trainer._call_callback_hooks("on_pool_start", *args, **kwargs)
+        self.trainer._call_lightning_module_hook("on_pool_start", *args, **kwargs)
+        self.trainer._call_strategy_hook("on_pool_start", *args, **kwargs)
+
+    def _on_evaluation_model_eval(self) -> None:
+        """Sets model to eval mode."""
+        self.trainer._call_lightning_module_hook("on_pool_model_eval")
+
+    def _on_evaluation_model_train(self) -> None:
+        """Sets model to train mode."""
+        self.trainer._call_lightning_module_hook("on_pool_model_train")
+
+    def _on_evaluation_end(self, *args: Any, **kwargs: Any) -> None:
+        """Runs ``on_{validation/test}_end`` hook."""
+        self.trainer._call_callback_hooks("on_pool_end", *args, **kwargs)
+        self.trainer._call_lightning_module_hook("on_pool_end", *args, **kwargs)
+        self.trainer._call_strategy_hook("on_pool_end", *args, **kwargs)
+
+        # reset the logger connector state
+        self.trainer._logger_connector.reset_results()
+
+    def _on_evaluation_epoch_start(self, *args: Any, **kwargs: Any) -> None:
+        """Runs ``on_epoch_start`` and ``on_pool_epoch_start`` hooks."""
+        self.trainer._logger_connector.on_epoch_start()
+        self.trainer._call_callback_hooks("on_epoch_start", *args, **kwargs)
+        self.trainer._call_lightning_module_hook("on_epoch_start", *args, **kwargs)
+
+        self.trainer._call_callback_hooks("on_pool_epoch_start", *args, **kwargs)
+        self.trainer._call_lightning_module_hook("on_pool_epoch_start", *args, **kwargs)
+
+    def _evaluation_epoch_end(self, outputs: List[EPOCH_OUTPUT]) -> None:
+        """Runs ``pool_epoch_end``"""
+        self.trainer._logger_connector._evaluation_epoch_end()
+
+        # with a single dataloader don't pass a 2D list
+        output_or_outputs: Union[EPOCH_OUTPUT, List[EPOCH_OUTPUT]] = (
+            outputs[0] if len(outputs) > 0 and self.num_dataloaders == 1 else outputs
+        )
+
+        # call the model epoch end
+        self.trainer._call_lightning_module_hook("pool_epoch_end", output_or_outputs)
+
+    def _on_evaluation_epoch_end(self) -> None:
+        """Runs ``on_pool_epoch_end`` hook."""
+        self.trainer._call_callback_hooks("on_pool_epoch_end")
+        self.trainer._call_lightning_module_hook("on_pool_epoch_end")
+
+        self.trainer._call_callback_hooks("on_epoch_end")
+        self.trainer._call_lightning_module_hook("on_epoch_end")
+        self.trainer._logger_connector.on_epoch_end()
